flask_blog_new/requirements.txt
alembic==1.0.11
backoff==1.8.0
bcrypt==3.1.7
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
Flask==1.0.3
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.12
MarkupSafe==1.1.1
Pillow==6.0.0
pycparser==2.19
python-dateutil==2.8.0
python-editor==1.0.4
requests==2.22.0
simple-geoip==0.1.0
six==1.12.0
SQLAlchemy==1.3.5
urllib3==1.25.3
Werkzeug==0.15.4
WTForms==2.2.1
